changed into floating data type

+++ main.py
@@ -17,14 +17,14 @@ operations = {
     "/" : division
 }
 def calculator():
-    num1 = int(input("What is the first number?: "))
+    num1 = float(input("What is the first number?: "))
     for symbol in operations:
         print(symbol)
     should_continue = True
 
     while should_continue:
         operation_symbol = input("Pick an operation: ")
-        num2 = int(input("What is the next number?: "))
+        num2 = float(input("What is the next number?: "))
         calculation_function = operations[operation_symbol]
         answer = calculation_function(num1, num2)
